requirements.txt
Package            Version
------------------ --------
annotated-types    0.7.0
anyio              4.4.0
certifi            2024.6.2
charset-normalizer 3.3.2
click              8.1.7
dnspython          2.6.1
email_validator    2.2.0
exceptiongroup     1.2.1
fastapi            0.111.0
fastapi-cli        0.0.4
h11                0.14.0
httpcore           1.0.5
httptools          0.6.1
httpx              0.27.0
idna               3.7
Jinja2             3.1.4
markdown-it-py     3.0.0
MarkupSafe         2.1.5
mdurl              0.1.2
orjson             3.10.6
pip                24.1.1
psycopg2           2.9.9
psycopg2-binary    2.9.9
pydantic           2.8.0
pydantic_core      2.20.0
Pygments           2.18.0
pyshortener        1.1.2
python-dotenv      1.0.1
python-multipart   0.0.9
PyYAML             6.0.1
requests           2.32.3
rich               13.7.1
setuptools         70.0.0
shellingham        1.5.4
sniffio            1.3.1
starlette          0.37.2
typer              0.12.3
typing_extensions  4.12.2
ujson              5.10.0
UNKNOWN            0.0.0
urllib3            2.2.2
uvicorn            0.30.1
uvloop             0.19.0
watchfiles         0.22.0
websockets         12.0
wheel              0.43.0
